+++ portfolio/src/main/webapp/index.html
@@ -9,6 +9,8 @@
     <script src="//ajax.googleapis.com/ajax/libs/jquery/1.7.2/jquery.min.js" type="text/javascript"></script>
     <script src="jquery_functions.js" type="module"></script>
     <script src="background_functions.js" type="module"></script>
+    <script src="helper_functions.js" type="module"></script>
+    <meta name="viewport" content="width=device-width,initial-scale=1.0">
   </head>
   <body>
       <div id="content">
@@ -33,9 +35,9 @@
         <h2>Click for a fun fact!</h2>
         <button onclick="displayFunFact()" class="transparent-button">FACT ME!</button>
         <div id="fact-container" class="fact-text"></div>
-        <h2> Click to change the background! </h2>
+        <h2 id = "background-image-header"> Click to change the background! </h2>
         <button class="transparent-button" id="change-background">CHANGE!</button>
-        <p> The background images are photos from some of the favorite places I've visited in Colorado. </p>
+        <p id = "background-image-description"> The background images are photos from some of the favorite places I've visited in Colorado. </p>
         <p id="background-caption">This image is of the summit of Mount Bierdstadt.</p>
         <div id="fact-container" class="fact-text"></div>
       </div>
